adding a builder for this logic.
